requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==5.8.0
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
